Backend/requirements.txt
fastapi
uvicorn
keras
pillow
numpy
